requirements.txt
pandas>=1.5.0
polars>=0.18.0
numpy>=1.22.0
scikit-learn>=1.2.0
scipy>=1.10.0

# Report Generation
reportlab>=4.0.0
python-pptx>=0.6.21

# LLM Integration
openai>=1.0.0
google-generativeai>=0.1.0

# Data I/O
openpyxl>=3.10.0
pyarrow>=12.0.0
sqlalchemy>=2.0.0

# Web API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Configuration
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
requests>=2.31.0
matplotlib>=3.7.0
Pillow>=9.0.0
